requirements.txt
APScheduler==3.8.1
apturl==0.5.2
arrow==1.2.1
asgiref==3.4.1
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifultable==0.8.0
binaryornot==0.4.4
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
Click==7.0
colorama==0.4.3
command-not-found==0.3
cookiecutter==1.7.3
cryptography==36.0.2
cssselect2==0.4.1
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.0
defer==1.0.6
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.2
django-admin-sortable2==1.0.2
django-angular==2.3
django-appconf==1.0.5
django-apscheduler==0.6.0
django-classy-tags==2.0.0
django-cms==3.9.0
django-crispy-forms==1.14.0
django-entangled==0.4
django-filer==2.1.2
django-formtools==2.3
django-fsm==2.8.0
django-fsm-admin==1.2.4
django-ipware==4.0.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-polymorphic==3.1.0
django-post-office==3.5.3
django-rest-auth==0.9.5
django-sekizai==2.0.0
django-select2==7.9.0
django-shop==1.2.4
django-treebeard==4.5.1
django-widget-tweaks==1.4.12
djangocms-admin-style==2.0.2
djangocms-cascade==1.3.7
djangocms-text-ckeditor==4.0.0
djangorestframework==3.12.4
duplicity==0.8.12.0
easy-thumbnails==2.8
entrypoints==0.3
fandogh-cli==1.26.10
fasteners==0.14.1
feedparser==6.0.8
filelock==3.4.0
future==0.18.2
html5lib==1.1
httplib2==0.14.0
idna==2.8
ipython==7.29.0
jedi==0.18.0
Jinja2==3.0.3
jinja2-time==0.2.0
jsonfield==3.1.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.6.4
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
package-name==0.1
paramiko==2.6.0
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2021.11.15
platformdirs==2.4.0
poyo==0.5.0
prompt-toolkit==3.0.22
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==1.9.73
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-slugify==5.0.2
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.4.1
reportlab==3.5.34
requests==2.26.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
retrying==1.3.3
SecretStorage==2.3.1
sgmllib3k==1.0.0
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
svglib==1.1.0
systemd-python==234
termgraph==0.5.3
text-unidecode==1.3
tinycss2==1.1.0
toml==0.10.2
traitlets==5.1.1
tzdata==2021.5
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.2
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
xkit==0.0.0
